Fix getOriginalROMMethod() crash indexing off the end of the J9Class VFT
When using Interpreter Profiling data to inline or devirtualize a call
it's possible for the isInstanceOf() check to pass despite the profiled
class not being compatible with the caller's context. This can happen
when compiling for AOT where the caller is sharing a J9ROMMethod with
a 2nd classloader and the callee in the bytecodes is an Abstract class
where some method is not defined in the Abstract class but is required
by an interface. This results in the the Abstract classes VFT containing
a reference to the Interface's J9Method. Because the VFT index for any
interface methods can differ between different implementors, if we use
the VFT index to retrieve a method from a profiled class that is from
an incorrect context several different bad things can happen. The
isInstanceOf() check can fail to protect us here because isInstanceOf()
will return true when checking the profiled class against an Interface
class retrieved from the Abstract classes VFT entry for an unimplemented
method of an Interface.

The fix is in two parts. 1) Inlining, 2) Codegen

1) There is an existing attempt to refine the _receiverClass for a
virtual CallSite that refers to an Interface class. This attempt
fails when compiling for AOT because the "optimizeForAOT" parameter
of the isInsterfaceOf() call and the "returnClassForAOT" parameter
of the getClassFromConstantPool() call was set to false. This fix adds
the required checks to make sure that setting the parameters to true is
safe for AOT compiles thereby allowing the refinement of the
_receiverClass from an interface into an Abstract class.

The fix also insures that any attempt use profile directed inlining when
the CallSite _receiverClass is an Interface will be aborted.

2) When code generator is attempting to perform a profile directed
devirtualization, this fix will check if the class of the call site
method is an Interface. In such cases the devirtualization attempt
will be aborted.

Signed-off-by: Kevin Langman <langman@ca.ibm.com>