Fix the JVM options passing to hangTest
The hang.sh script expects the second argument to be
a classpath. If JVM options (such as -XX:+UseJITServer)
are specified, the test will not run.

This commit fixes the issue by moving $JVM_OPTIONS to
a later position in the playlist.xml and modifying hang.sh
to take the options as a third argument.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>