Specify VFT slot for InterpreterEmulator refined callee of virtualCall_*
By providing the vTable slot here we make sure to get the correct value
in the TR_J9VirtualCallSite created in visitInvokestatic().

Previously the slot was defaulted to 0. Because it's expected to be an
interpreter vTable offset (from the beginning of the J9Class), this
could resulted in a crash within TR_J9VirtualCallSite due to loading the
J9Class eyecatcher where a J9Method pointer was expected.