Use correct loop condition when searching for methods
As of JDK8 we permitted the addition of methods when redefining classes.
This means that between two versions of the same class, the number of
methods may not be the same. As a result we can not use the same
romMethodCount when iterating ramMethods for different versions of the
same class.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>