Fix overflow in Balanced expand
Total heap resizing (based on GC overhead heuristics) may suggest to
expand to a smaller value than current total heap size, leading to
contraction (what might not be problem by itself). The bigger problem is
that internally it's represented as a big (overflowed) unsigned integer
and while arithmetic (+/- operators) will silently work (we will
successfully contract), that's not the case for comparisons (<,>
operators).

If there was a large object alloc request and there is not free enough
memory we may want to expand, but since this alloc size is likely to be
less than the fake expand, it won't be able to trigger the real expand.

Eventually, this may lead to failing to allocate and OOM condition.

The fix is simply to prevent the overflow math and set the expansion
value to 0, if calculated value is negative.

Fixes: https://github.com/eclipse-openj9/openj9/issues/13584

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>