Prevent recompilation for failed compilations retried at lower opt levels
Under certains conditions, a compilation that fails at a particular optimization
level can be retried at a lower optimization level, hoping that the compilation
will pass. However, in this case we must prevent future recompilation at a
higher opt level, because this could lead to a perpetual compilation process.
Example:
 warm compile
 hot recompile; fail
 warm recompile
 hot recompile; fail
 ...

This commit disables GCR (guarded counting recompilation) and EDO (exception directed
optimization) for compilation retrials to avoid the possibility of the method being
upgraded to a higher optimization level in the future. Note that sampling is already
disabled, for the same reason.

Fixes #13548

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>