Remove useless 'thrtrace' module
The native provided is for `com.ibm.oti.vm.thread.Tracing` which is
not included in an SDK, neither is the shared library.

Even if it were, it wouldn't provide anything useful because
`OMR_THR_TRACING` is not defined. If it were there would be a link
error because `omrthread_reset_tracing()` is not defined by OMR.

Remove dead code related to `OMR_THR_TRACING`.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>