Modify JITServerTest to use a new port for each subtest
JITServer tests run multiple test cases where servers frequently
start and shut down. If the port is not immediately released
back into the system, or another process grabs it before the next
server can start, all subsequent test cases will fail because server
will not be able to start.

This commit modifies the test to generate a new JITServer port
for each test case, to minimize the chances that the port is not
available.

Closes: #13801

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>