Do not enter the write mutex after the read mutex
It is possible that we will enter write mutex inside
updateRuntimeFullFlags(). If it is called from a reader, this could
result in a deadlock if there is another thread waiting for the reader
to finish and trying to lock the cache. 

1. For readers, move the call of updateRuntimeFullFlags() at the
beginning of enterReaderMutex() before the reader count is incremented. 

2. Add assertions that write mutex are not acquired after read mutex.

Closes #13837

Signed-off-by: Hang Shao <hangshao@ca.ibm.com>