Only fail when an existing env value is trying to be modified
If the new value is identical to the old value just ignore the new value
and continue on; otherwise collect the invalid keys and fail the
restore.

Implementation note:  A String.format call is used instead of
System.err.printf because printf was producing strange newlines in the
output for each %s substitution.  It was not determined why that was the
case.

Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>