Support inline allocation for value types
Add requireHelperCallValueTypeAllocation to check if inline allocation
is allowed when evaluating the "new" opcode and when value
type is enabled. If "new jitNewValue" is used to create a non
value type, or if "new jitNewObject" is used to create a value type,
InstantiationError exception needs to be thrown. Use helper call in
these cases.

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>