InjectedInvoker behaviour modified in JDK18
In JDK17, InjectedInvoker does not have the NESTMATE attribute. In JDK18,
InjectedInvoker's definition has been updated, and it is given the NESTMATE
attribute. The new changes invalidate the filter that is used to identify
InjectedInvoker classes for modifying its class name. The filter has been
updated to work with JDK18.

Closes: https://github.com/eclipse-openj9/openj9/issues/13878

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>