Don't override counts if specified by user in subset
There's a corner case where specifying counts in an option set while the
SCC is enabled results in the option set getting ignored. The reason
this happens is because in jitHookInitializeSendTarget, the update is
guarded by TR::Options::getCountsAreProvidedByUser, which only returns
true when counts set by the user outside an option set. This commit
fixes this issue.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>