Remember JITServer AOT cache records when writing AOT relocation data
For each client SCC offset written into AOT relocation data during
out-of-process AOT compilations, we remember the corresponding AOT
cache record (creating it on demand if necessary) and the offset
into AOT relocation data where the SCC offset is stored. The list of
<AOT cache record, relocation data offset> pairs is stored in the
compilation object. This list will be used to store the serialized
AOT method in the JITServer AOT cache at the end of the compilation.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>