Correctly update sequence number when client aborts compilation early
The code previously incorrectly assumed that the only case when the
client can abort the compilation early is when it replies to the message
requesting the whole CHTable, at which point the last processed critical
sequence number is not yet updated in the client session.

Aborts can also occur when requesting uncached classes or VMInfo from
the client, which can happen after the sequence number is already
updated. To distinguish these two cases, we add a flag to keep track
of whether the sequence number was already updated or not.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>