JITServer AOT cache load implementation
This commit adds the code to serve serialized methods from the JITServer
AOT cache in response to AOT compilation requests and load these methods
at the client after deserializing them and storing in the local SCC.

If the AOT deserialization or load fails, the next attempt to compile
the method will bypass the JITServer AOT cache.

AOT serialization record IDs newly added to the client's cache during
deserialization are sent to the server with subsequent compilation
requests and are saved in the client session at the server.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>