Avoid scanning the compilation queue for duplicates
When a compilation request is added to the compilation queue,
the JIT needs to scan the queue for a duplicate request and
coalesce the two requests into one if a duplicate exists.
In some corner cases the compilation queue can grow to thousand
of entries and scanning the queue for duplicates adds a lot of
overhead. This commit avoids scanning the queue based on the fact
that asynchronous queued entries always have
j9method->extra == J9_JIT_QUEUED_FOR_COMPILATION
This optimization can be disabled by defining the following env var:
TR_DisableSkipSearchingForRequestDuplicates = 1

Partially addresses issue #13807

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>