Impose a limit on the compilation queue size
In corner cases the compilation queue size can grow to tens of
thousand of entries. A large compilation queue adds overhead in
terms of memory and CPU (inserting an entry requires queue scanning
to find the proper place). To avoid these overheads, this commit
imposes a limit for the compilation queue size. When the limit is
reached, future asynchronous compilation requests for interpreted
methods are rejected, and the invocation count for the methods in
question is replenished. The rejected methods will continue to
run interpreted until their invocation count reaches 0 again.
Postponing such compilation requests is not expected to be detrimental
to performance because those methods would have waited a very
long time in the queue anyway.
This compilation queue size limit does not apply to recompilation
requests, to JNI methods (which don't use an invocation count) or
to synchronous requests which need to be processed as soon as possible.
The default value for the compilation queue size limit is set to 5000.
This value can be changed with -Xjit:compilationQueueSizeLimit=<NNN>
This feature can be disabled by setting a very large value for the
compilation queue size limit.

Partially addresses issue #13807

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>