Make the JIT use a single consistent vmentry for each MethodHandle
...when OpenJDK MethodHandles are enabled.

When the JIT processes a call to invokeBasic() on a known MethodHandle,
mh, it loads mh.form.vmentry to find the MemberName that specifies the
LambdaForm-generated method that implements mh. It then refines the call
into a direct call to the implementing method. Typically this call will
be inlined after refinement, but it may be generated as a call instead,
e.g. due to disableInlining, or dontInline={...}, etc. Either way, the
generated code may rely on the implementing method remaining loaded. For
example, inlined code may need to resolve constant pool entries, load
from pseudo-TOC entries that get released on unload, or OSR to continue
execution in the interpreter.

The MemberName holds a reference to the generated class, so the
implementing method will stay loaded as long as the MemberName is live.
The references from the MethodHandle to the LambdaForm and then from
there to the MemberName are usually enough to hold the MemberName live
as long as the MethodHandle, so that the generated code works when the
invokeBasic() call is reached at runtime. However, it is possible for
some of these references to change after (or even during) compilation.
In particular:

1. A MethodHandle may have its LambdaForm replaced with another
   semantically equivalent one.

2. A LambdaForm may be redundantly compiled to bytecode multiple times
   with no synchronization to prevent multiple distinct MemberName
   objects from being observed. The generated classes are only nominally
   different, so they're generally interchangeable.

If either of these situations arises after the JIT compiler has observed
the vmentry for a particular MethodHandle, the update can sometimes
make the original vmentry unreachable, and if it does, that can allow
the generated bytecode to be unloaded even though it's still needed for
a caller's JIT-compiled code.

This commit adds a field to MethodHandle called jitVMEntryKeepAlive. The
first time the JIT compiler observes the vmentry corresponding to a
given MethodHandle, it updates the field to point directly from the
MethodHandle to the observed MemberName instance. Whenever the JIT
subsequently looks for the vmentry of that MethodHandle, it will simply
use the reference from jitVMEntryKeepAlive, ignoring form.vmentry. In
this way, the compiler ensures that the MemberName it observes will live
at least as long as the MethodHandle, and therefore so will the
generated bytecode. The bytecode can still be unloaded once the
MethodHandle is collected, but at that point any invokeBasic() call that
was transformed for the handle had better already be unreachable.

Note that it's fine to miss the updates (1) and (2). For (1), the new
LambdaForm is required to be equivalent to the old, and threads are
allowed to keep running the old one indefinitely. In the case of (2),
for a given LambdaForm, any generated class is as good as any other.

Fixes #13588