Avoid folding MethodHandle.form, which is mutable
I'm not aware of any actual code miscompiled due to the compiler folding
loads of MethodHandle.form to a known object. This is a precaution to
prevent any such incorrect folding in the future.

The compiler still does dereference MethodHandle.form to determine which
generated method implements a given MethodHandle. The generated code
assumes that the implementing method doesn't change, which is allowed,
and which will not cause the same bizarre kinds of behaviour as
incorrect known object information.