Switch to AOT compilations to take advantage of JITServer AOT cache
In the current implementation, in order for JITServer to send a
serialized AOT body from its cache, the compilation at the client
must be an AOT compilation. Thus, currently, if we want to maximize
the usage of the JITServer AOT cache we must use -Xaot:forceaot at
the client.
This commit eliminates this restriction. If -XX:JITServerUseAOTCache
option is present at the client, the client will try to generate as
much AOT as possible, but still keep the cheap compilations local
in order to avoid network latency.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>