Add Xjit:smallMethodBytecodeSizeThresholdForJITServerAOTCache option
Generating too many AOT bodies can be detrimental for rampup due to
the GCR recompilation mechanism which can add a lot of runtime
overhead once all those bodies start counting.
This commit limits the number of AOT compilations when JITServer
AOT caching is used. The idea is that small methods may take little
time to compile and we don't have to use the JITServer AOT cache
for them because the savings would be too small. Therefore we can
afford to JIT compile them remotely.
Which methods are small enough is determined by the newly introduced
option -Xjit:smallMethodBytecodeSizeThresholdForJITServerAOTCache=<N>
The default value for this option is 0, which means disabled.
Experiments have shown that, in some cases, small rampup improvements
can be had with this option set to 32.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>