Turn off GCR counting sooner if many GCR compilations have been queued
OpenJ9 turns off counting for GCR (counted guarded recompilation)
if too many GCR induced compilation requests are queued. The problem
with the current implementation is that the decision to turn off
counting is done every 0.5 seconds, and in half a second very many
GCR induced recompilations can be queued.
This commit changes the granularity of checking whether GCR counting
should be turned off to the period of the sampling thread, which is
typically around 10 ms.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>