Add wrapper around rtlog printf
This commit adds wrappers around the JITRT_PRINTF and related macros to
check if the thread should use the global rtLog or the per compilation
thread rtLog. If a non-compilation thread happens to be running any code
that wants to print to the rtLog, it should print to the global log.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>