Process aladd offsets that can fit in 32 bits
New initialization is enabled on 64-bit compressed refs in addition to
32-bit. Array accesses use aladd instead of aiadd on 64-bit compressed
refs and so the code that examines array accesses should check for a
long constant offset (in addition to int constant offset as is the case
presently). This would allow the code to engage and attempt
transformations on 64-bit compressed refs for arrays. Since the rest of
the data structures are using 32-bit offsets we restrict the actual
value of the long constant to be in the 32-bit range since we do not
believe this to be a significant limiting factor in practice.

Signed-off-by: Vijay Sundaresan vijaysun@ca.ibm.com