Set initial heap size equal to Xmx value under -Xtune:throughput
Experiments have shown that there is a non-negligible throughput
improvement of applications (3-5%) when Xms==Xmx. Thus, if the
user specifies -Xtune:throughput on the command line, but does
not specify an Xms value, then the JVM will internally set the
initial heap size (Xms value) to the maximum heap size (Xmx value).

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>