Enable profile directed guarded devirtualization by default for JIT on Z
P and X do not check the 'enableProfiledDevirtualization' option before
performing profiled directed guarded devirtualization. This patch
removes this check in Z codegen to bring it in line with P and X, and
adds a check on the profiled virtual method to ensure it is not
interpreted (also in line with P and X).

Fixes: #13606
Signed-off-by: Spencer Comin <spencer.comin@ibm.com>