Change JIT heuristics/options under -Xtune:throughput and other cleanup
When -Xtune:throughput is specified the JIT will internally
set some options that favor throughput at the expense of
footprint, CPU or start-up time. The options that are internally
set can still be overridden with -Xjit options.

This commit also eliminates the JIT parsing for the "-client" and
"-server" options since it wasn't working anyway.
Another change is that only the last appearance of "-Xquickstart",
"-Xtune:virtualized" and "-Xtune:throughput" will take affect
(this is ensured by VM option parsing and setting of a corresponding
flag). Previously, "-Xquickstart" had precedence over "-Xtune:virtualized"
if both options appeared on the command line.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>