Fix native.encoding and properly initialize the console charset
See JEP 400 https://openjdk.java.net/jeps/400

native.encoding should always be the platform encoding and is not
affected by values set on the command line. Even on jdk18+ it remains
the platform encoding while file.encoding defaults to UTF-8.

In jdk18+ the console charset/encoding needs to use the native.encoding
and not the default (UTF-8) charset.

Signed-off-by: Peter Shipton <Peter_Shipton@ca.ibm.com>