Do unprivileged CRIU dump if requested+supported
If the user requests an unprivileged CRIU dump we check
if a non-NULL `criu_set_unprivileged` symbol can be found via
dlsym() and if so call it with `true` as the parameter.

If the user requests an unprivileged CRIU dump but the symbol
cannot be found we throw an exception.

If the user does not request an unprivileged CRIU dump we do
nothing as CRIU (currently) defaults to privileged.

Also link against libdl to get access to dlsym().

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>