Modify initial invocation counts at high occupancy
To relieve some of the pressure on the code cache during periods of high
cache occupancy, the initial invocation counts are increased during
those periods. Transitions between low and high cache occupancy are
detected in samplerThreadProc; when such a transition takes place, the
initial invocation counts for subsequent new methods are set by
jitHookInitializeSendTarget to a configurable value.

Options controlling the behaviour of the high cache occupancy code are:

- OMR::Options::_highCodeCachePercentage in omr, which defines the
  threshold at which the code cache is considered to have high occupancy
- OMR::Options::_highOccupancyInitialCount and
  OMR::Options::_highOccupancyInitialBCount in omr, which store the
  initial invocation counts to be used at high occupancy

Related: #14299
Signed-off-by: Christian Despres <despresc@ibm.com>