Option to print JITServer malloc statistics at JIT dump on Linux
When the environment variable TR_PrintJITServerMallocStats is set, use
the malloc_stats() function to print the malloc heap statistics (such
as the amount of allocated and reserved space in each per-thread heap)
when a JIT dump is triggered, on both the server and the client.

This data is useful for profiling JITServer memory usage since it
currently uses the malloc heap heavily (e.g. in message serialization).

Note that this feature is only supported on Linux since malloc_stats()
is a non-standard glibc extension.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>