Extract relocateAndRegisterThunk() from TR_RelocationRecordThunks
This makes the J2I thunk loading and registration logic available to be
called for the J2I thunk from method SVM validation that will be defined
in a subsequent commit.

The calls to pthread_jit_write_protect_np() are removed because although
this method is writing to the code cache, the write protection is
already handled in applyRelocationAtAllOffsets().