Replace forceUnresolvedDispatch() with isResolved*DispatchGuaranteed()
Certain transformations in the compiler rely on a guarantee that a given
resolved call node will simply call the target method without attempting
to resolve a constant pool entry, since the constant pool entry may not
correspond to the call in the usually expected way.

We always have this guarantee in non-AOT compilations.

In AOT compilations, we currently have this guarantee for direct calls
when using the symbol validation manager (SVM) on x86, Power, and z.

In the SVM case, each code generator must now opt in to this class of
transformations by explicitly declaring that it provides the needed
guarantee for direct calls, virtual calls, or both.

Special-case virtual and interface resolution in AOT now respects
isResolved*DispatchGuaranteed(). In particular, private invokevirtual,
private invokeinterface, and final Object invokeinterface will all
become direct, so heed isResolvedDirectDispatchGuaranteed(), and
non-final Object invokeinterface will become virtual, so now heeds
isResolvedVirtualDispatchGuaranteed().

Additionally, refinement of invokeBasic() and linkTo*() now also
respects these queries, though it shouldn't (currently) be attempted in
AOT regardless. Refinement requires a known object index for either the
MethodHandle or the MemberName, and AOT does not support the known
object table.

It probably also makes sense to gate value propagation's type refinement
of indirect calls on isResolvedDirectDispatchGuaranteed() for fixed type
and isResolvedVirtualDispatchGuaranteed() for type bounds, but this is
not done in this commit. That change would require modifications to OMR,
and there is currently no problem with call refinement in VP because it
respects canDevirtualizeDispatch(), which is false in AOT compilations.

As a result of the above, this commit effectively stops AOT compilations
using the SVM from generating resolved virtual calls for invokeinterface
calling non-final methods of Object. Instead, such calls are left to be
treated as (unresolved) interface calls as in non-SVM AOT. This ilgen
transformation can be re-enabled for a particular code generator by
overriding guaranteesResolvedVirtualDispatchForSVM() to return true, but
first it will be necessary to ensure that the code generator really
provides that guarantee. If the sequence generated for a resolved
virtual call is simply a call through the vtable, then providing the
guarantee should only involve:
- sending execution to the resolved path when the SVM is enabled, and
- generating an appropriate J2I thunk validation.

However, the required J2I thunk validation does not yet exist. It will
be implemented separately.