PGC consumption rate
Current rate is not quite correct since
1) it was not accounting for heap resizing changes, while it should
2) and accounting for sweep reclamation, while it should not

As a result of (1), during heap expansion (that is initially justified,
like ramp-up phase) we would underestimate heap consumption, what would
delay GMP kickoff, and in some cases we would be forced to expand even
more to be able to continue PGC. That could create positive feedback
where we would continue expanding till fully expanded.

The rate formula is completely changed to rely simply on current PGC
metrics:
consumed (survived) regions - freed (evacuated) regions

...as opposed to measuring the difference in heap occupancy between
previous and current GC.

In a stable state, those 2 approaches are very similar, but during heap
resizing, the new one is far more correct.

As a side effect of this change, GMP cycles during heap expansion will
be more frequent, but in a way it is good as it will help with more
precise heuristics (like live set estimate).

defragmentRegionConsumptionRate suffers from the same problem, but it
will be scrutinized separately, since it affects a different huristic.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>