Correctly throw UOE from setSecurityManager
In some cases, warning messages are no longer being printed when the input
argument is null. Revising the fix in #14329 to correctly throw
UnsupportedOperationException (UOE) in case of a null input argument when the
security manager is disallowed.

When the security manager is not allowed to be set dynamically, return if the
input argument is null. UnsupportedOperationException should only be thrown for
a non-null input argument.

Fixes: #14445

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>