Fix for increased tlhSurvivorDiscardThreshold for PGC
Fix for object allocate in copy path for CopyForward in PGC, when
allocated TLH copy cache is not large enough for object requested (what
may cause an assert later).

We will now retry the allocate, but we will force single object
allocation path.

The failure could really occur only if tlhSurvivorDiscardThreshold is
increased above _minCacheSize (what defaults to tlhMinimumSize).

Default tlhSurvivorDiscardThreshold may need to be increased in future
(or through manual tuning) to prevent too frequent single object
allocation (and contention caused by it).

A bit of cleanup (UDATA to uintptr_t) in all 'reserveMemory' methods,
but there is more cleanup in this very large file.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>