Eliminate the effect -Xaggressive has on switching to profiling
The option `-Xaggressive` allows switching to profiling of hot compilations
in a more aggressive manner. This has been seen to have negative performance
effects on some applications.
This commit eliminates the effect `-Xaggressive` has on switching to profiling.
More aggressive switching to profiling can be enabled with the newly created
option `-Xjit:aggressiveSwitchingToProfiling`

Another change is the elimination of code that tried to disable downgrading
to cold optimization level. That code had no effect because it was conditional
on an option that was not yet set at that point.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>