Replaces uses of currentThread with walkThread in jswalk
There are cases where a non JVM thread may trigger a javacore. In these
cases the currentThread will be NULL. This will cause crashes when we
need to walk certain jit frames as the currentThread is used to
facilitate this. This PR replaces uses of currentThread with walkThread
in these cases.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>