Avoid r11 in iTable offset tag bits check on Power
We need to maintain the contents of `r11` for the case where we have
to call the helper. The GPR result of the `andi.` therefore needs to go
into another register. The GPR result isn't actually needed, only the
result of the test against 0 (stored in `cr0`) is, so we can use any
register that's free at that moment. `r0` is a good candidate.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>
(cherry picked from commit c5de65dfa82cd3d8b6b993725b566a1ad2420d7d)