Delay lock-related operations in single thread mode
This PR delays notify/notifyAll operations by intercepting the
Object.notify* INLs and saving the instances and operations to a queue.
These instances must be saved as global JNI refs as we cannot save them
in localref storage and disallowing a GC is not feasible. At the end of
the single thread phase all the operations will be executed by the
checkpointing thread.

See https://github.com/eclipse-openj9/openj9/issues/14584

Handles case 4) Notify

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>