Improve checkcast/instanceof performance for interfaces on x86/x64
PR #2361 introduced a number of JIT performance enhancements for checkcasts
involving interface classes.  Two of the most significant changes were the
introduction of a one-slot, dynamic cache to cache the last successful instance
class that matched the cast class.  It also inlined the itable walk to determine
if the interface class implemented the cast class.

Each time a successful checkcast was performed the cache would be updated with
that result at runtime (it was an LRU cache).  This cache was implemented as a
single, 8-byte data snippet located in the code cache.  The performance of this
approach may be acceptable for checkcast sites that do not see more than one
instance class, or when there is only a single thread of execution along this
path.  However, in the presence of multiple threads and multiple instance classes
the performance penalty of multiple threads writing to the same data address (the
class cache) is quite significant due to hardware cache coherency protocols.
In fact, the performance overhead gets significantly worse the more threads that
are involved.

There is no information provided in #2361 for what motivated the change or the
workload it was expected to benefit.

This PR modifies that implementation as follows:

* Consult available profiling information to determine how many instance classes
  may be seen by this checkcast/instanceof site.  If there are more than one then
  do not use a cache.  If there is no information available then do not use a
  cache as the characteristics are unknown.

* If there is a single profiled instance class then pre-populate the cache with
  that class.  Do not update the cache at runtime.

* If a cache is not used, then inline the interface table walk in mainline code
  rather than from outlined instructions.

Three environment variables have been introduced to control behaviour of this
evaluator:

1) `TR_updateInterfaceCheckCastCacheSlot` : when set the interface cast cache
   slot will be updated at runtime (i.e., this is the original behaviour)

2) `TR_forceDisableInterfaceCastCache` : never use the cache regardless of
   profiling information.

3) `TR_forceEnableInterfaceCastCache` : force the use of the cache regardless of
   profiling information.

2) and 3) are mutually exclusive.  Behaviour is undefined if both are set.

Using 1) and 3) will restore original behaviour.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>