Simplify -Xaggressive code
One of the effects of the `-Xaggressive` option is to enable static final field
folding for the JCL classes. Static final field folding is already enabled
under OSR (which is on by default) so this code is now obsolete and will be
deleted in this commit.
For the cases where OSR is disabled we can still enable static final field
folding for JCL classes with the `TR_EnableJCLStaticFinalFieldFolding`
environment variable.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>